replace more ints with int

+++ R/utils-navigate-nest.R
@@ -5,24 +5,24 @@
 #' @importFrom rlang seq2
 #' @keywords internal
 next_non_comment <- function(pd, pos) {
-  if (length(pos) < 1 || is.na(pos) || pos >= nrow(pd)) {
-    return(integer(0))
+  if (length(pos) < 1L || is.na(pos) || pos >= nrow(pd)) {
+    return(integer(0L))
   }
   candidates <- seq2(pos + 1L, nrow(pd))
   if (all(candidates %in% which(pd$token == "COMMENT"))) {
-    return(integer(0))
+    return(integer(0L))
   }
   setdiff(candidates, which(pd$token == "COMMENT"))[1]
 }
 
 #' @rdname next_non_comment
 previous_non_comment <- function(pd, pos) {
-  if (length(pos) < 1 || is.na(pos) || pos > nrow(pd)) {
-    return(integer(0))
+  if (length(pos) < 1L || is.na(pos) || pos > nrow(pd)) {
+    return(integer(0L))
   }
   candidates <- seq2(1L, pos - 1L)
   if (all(candidates %in% which(pd$token == "COMMENT"))) {
-    return(integer(0))
+    return(integer(0L))
   }
   last(setdiff(candidates, which(pd$token == "COMMENT")))
 }
@@ -59,14 +59,14 @@ next_terminal <- function(pd,
                           stack = FALSE,
                           vars = c("pos_id", "token", "text"),
                           tokens_exclude = c()) {
-  pd$position <- seq2(1, nrow(pd))
+  pd$position <- seq2(1L, nrow(pd))
   pd <- pd[!(pd$token %in% tokens_exclude), ]
-  if (pd$terminal[1]) {
+  if (pd$terminal[1L]) {
     pd[1, c("position", vars)]
   } else {
-    current <- next_terminal(pd$child[[1]], stack = stack, vars = vars, tokens_exclude = tokens_exclude)
+    current <- next_terminal(pd$child[[1L]], stack = stack, vars = vars, tokens_exclude = tokens_exclude)
     if (stack) {
-      bind_rows(pd[1, c("position", vars)], current)
+      bind_rows(pd[1L, c("position", vars)], current)
     } else {
       current
     }
@@ -83,7 +83,7 @@ extend_if_comment <- function(pd, pos) {
   if (pos == nrow(pd)) {
     return(pos)
   }
-  if (pd$token[pos + 1] == "COMMENT") {
+  if (pd$token[pos + 1L] == "COMMENT") {
     extend_if_comment(pd, pos + 1L)
   } else {
     pos
